Cancel transition when no matching choice
The original logic was separately checking if 1) there were any valid
defined transitions for the current state, and 2) if any of the
transitions' conditions matched. By doing this separately, the check
for whether an event transition was valid would provide incorrect
results when the only transitions that had conditions that all passed
were not defined transitions from the *current* state.

Once that check returned an incorrect result, the call later down the
line to pull out the supposed "matching" transition would come up with
air (or `nil`, in this case).